+++ src/middlewares/authentication.ts
@@ -13,6 +13,15 @@ export default async (req: Request, res: Response, next: NextFunction) => {
             state: req.originalUrl
         });
 
+        if (process.env.PROJECT_ID === undefined) {
+            throw new Error('Set environment variable `PROJECT_ID`');
+        }
+
+        req.project = {
+            typeOf: 'Project',
+            id: process.env.PROJECT_ID
+        };
+
         if (!req.user.isAuthenticated()) {
             // ログインページへリダイレクト
             res.redirect(req.user.generateAuthUrl());